requirements.txt
fastapi==0.70.0
pydantic==1.8.2
uvicorn==0.15.0
requests==2.26.0
pandas==1.3.3
torch==1.9.0
transformers==4.11.3
scikit-learn==0.24.2
vncorenlp==1.1.1
google-generativeai==0.1.0